Remove profile setting of ispyb.env user property
Remove the setting of the ispyb.env user property from profiles because
the ispyb.env-* profiles are expected to be activated only by the
ispyb.env system property.  Keeping the old behavior of profiles setting
ispyb.env as a user property would leave the door open for confusing
scenarios (e.g., the system property being set to a different value
than the user property of the same name if an ispyb.env-* profile were
explicitly enabled).