Remove profile setting of ispyb.site user property
Remove the setting of the ispyb.site user property from profiles because
the ispyb.site-* profiles are expected to be activated only by the
ispyb.site system property.  Keeping the old behavior of profiles
setting ispyb.site as a user property would leave the door open for
confusing scenarios (e.g., the system property being set to a different
value than the user property of the same name if an ispyb.site-* profile
were explicitly enabled).