Add REST web service property-based authentication
Add a RESTful web service property-based authentication module that
makes it easy to define users in an ISPyB property and eliminates
the need in development and testing for an LDAP server in order to
authenticate RESTful web services.  It can also be used for the GENERIC
site, or a new site, to get up and running more quickly for development
or for demoing without requiring an LDAP server.