Make ispyb-ejb GENERIC site properties generic
Change the ispyb-ejb properties for the GENERIC site to have generic
values and not reference the ESRF, ESRF-specific file system paths, ESRF
email addresses, etc.  This makes it easier to build and run the ISPyB
web app out of the box for the GENERIC site without having to change
various properties.